+++ Editor/NFTs_Details_Editor.cs
@@ -5,29 +5,28 @@ namespace NFTPort.Editor
     using UnityEditor;
     using Internal;
 
-    [CustomEditor(typeof(Mint_Custom))]
-    public class Mint_Custom_Editor : Editor
+    [CustomEditor(typeof(NFT_Details))]
+    public class NFTs_Details_Editor : Editor
     {
         public override void OnInspectorGUI()
         {
             
-            Mint_Custom myScript = (Mint_Custom)target;
+            NFT_Details myScript = (NFT_Details)target;
             
             
-            Texture banner = Resources.Load<Texture>("c_productmint");
+            Texture banner = Resources.Load<Texture>("c_nftdata_details");
             GUILayout.BeginHorizontal();
             GUILayout.Box(banner);
             GUILayout.EndHorizontal();
 
-            if (GUILayout.Button("Mint Custom NFT", GUILayout.Height(45)))
+            if (GUILayout.Button("GET NFT Details", GUILayout.Height(45)))
             {
                 PortUser.SetFromEditorWin();
                 myScript.Run();
             }
-        
 
             if(GUILayout.Button("View Documentation", GUILayout.Height(25)))
-                Application.OpenURL(PortConstants.Docs_Mint_Custom);
+                Application.OpenURL(PortConstants.NFTs_OfAccount);
             DrawDefaultInspector();
         }
     }