+++ Runtime/Internal/PortConstants.cs
@@ -8,13 +8,16 @@ namespace NFTPort.Internal
         public const string Docs_GettingStarted = "https://docs.nftport.xyz/docs/nftport/ZG9jOjUwMzc2NzA5-getting-started"; 
         public const string Docs_StorageFile = "https://docs.nftport.xyz/docs/nftport/ZG9jOjYwODM0NTY3-storage-upload-file";
         public const string Docs_StorageMetadata= "https://docs.nftport.xyz/docs/nftport/ZG9jOjYwODM0NTY4-storage-upload-metadata";
+        public const string Docs_NFTDetails = "https://docs.nftport.xyz/docs/nftport/ZG9jOjU3MTU2NTE5-player-wallet-connect";
         public const string NFTs_OfAccount = "https://docs.nftport.xyz/docs/nftport/ZG9jOjUyMzI4NTkz-nf-ts-owned-by-an-account";
         public const string NFTs_OfContract = "https://docs.nftport.xyz/docs/nftport/ZG9jOjUzNjI2MzQ0-nf-ts-of-a-contract-collection";
         public const string Docs_Mint_URL = "https://docs.nftport.xyz/docs/nftport/ZG9jOjU1MDM4OTgw-minting-w-url";
         public const string Docs_Mint_Custom = "https://docs.nftport.xyz/docs/nftport/ZG9jOjYzMDIzNDgx-customizable-minting";
         public const string Docs_AssetDownloader = "https://docs.nftport.xyz/docs/nftport/ZG9jOjU2NjAzOTE0-asset-downloader";
         public const string Docs_ConnectUserWallet = "https://docs.nftport.xyz/docs/nftport/ZG9jOjU3MTU2NTE5-player-wallet-connect";
-        public const string Docs_DeployContract = "https://docs.nftport.xyz/docs/nftport/b3A6MjE0MDYzNzU-deploy-a-contract-for-nft-products";
+        public const string Docs_DeployContract = "https://docs.nftport.xyz/docs/nftport/ZG9jOjYzOTc1Mzgw-deploy-contract";
+        
+        public const string Docs_Txns_Account = "https://docs.nftport.xyz/docs/nftport/ZG9jOjYzOTc1Mzgw-deploy-contract";
 
 
         public const string DiscordInvite = "https://discord.gg/w92sXkNmBR";
@@ -24,6 +27,7 @@ namespace NFTPort.Internal
         public const string BaseComponentMenu = "NFTPort/";
         public const string BaseFeatureSpawnerMenu = "NFTPort/Spawner/";
         public const string FeatureName_ConnectUserWallet = "Connect Player Wallet";
+        public const string FeatureName_NFT_Details = "NFT Data | NFT Details";
         public const string FeatureName_NFTs_OfContract = "NFT Data | of a contract-collection";
         public const string FeatureName_NFTs_OfAccount = "NFT Data | of an Account";
         public const string FeatureName_StorageFiles = "Storage | Upload File";
@@ -32,6 +36,7 @@ namespace NFTPort.Internal
         public const string FeatureName_Mint_Custom = "Mint | Customizable minting";
         public const string FeatureName_Mint_URL = "Mint | Easy mint using URL";
         public const string FeatureName_Deploy = "Deploy Contract";
+        public const string FeatureName_Txn_Account = "Transaction Data | Account";
 
     }
 }